科室的分页，权限的校验

+++ module-core/src/main/java/com/mujugroup/core/service/impl/DepartmentServiceImpl.java
@@ -1,5 +1,7 @@
 package com.mujugroup.core.service.impl;
 
+import com.lveqia.cloud.common.config.CoreConfig;
+import com.lveqia.cloud.common.exception.DataException;
 import com.lveqia.cloud.common.exception.ParamException;
 import com.lveqia.cloud.common.objeck.DBMap;
 import com.lveqia.cloud.common.util.StringUtil;
@@ -8,15 +10,19 @@ import com.mujugroup.core.mapper.HospitalMapper;
 import com.mujugroup.core.model.Department;
 import com.mujugroup.core.model.Hospital;
 import com.mujugroup.core.objeck.vo.department.DepartmentVo;
+import com.mujugroup.core.objeck.vo.department.ListVo;
 import com.mujugroup.core.objeck.vo.department.PutVo;
 import com.mujugroup.core.objeck.vo.SelectVO;
+import com.mujugroup.core.service.AuthDataService;
 import com.mujugroup.core.service.DepartmentService;
 import ma.glasnost.orika.MapperFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 
 @Service("departmentService")
@@ -24,20 +30,33 @@ public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentMapper departmentMapper;
     private final MapperFactory mapperFactory;
     private final HospitalMapper hospitalMapper;
+    private final AuthDataService authDataService;
 
     @Override
-    public boolean delete(String id) throws ParamException {
+    public boolean delete(int uid, String id) throws ParamException, DataException {
+        Map<String, String> map = authDataService.getAuthDataByUid(uid);
+        if (map.size() == 0) throw new DataException("当前用户没有数据权限,请联系管理员");
+        if (!map.containsKey(CoreConfig.AUTH_DATA_ALL)) throw new DataException("当前用户无最高数据权限，暂无法进行删除操作!");
         if (StringUtil.isEmpty(id)) throw new ParamException("科室编号不能为空");
         if (!StringUtil.isNumeric(id)) throw new ParamException("科室编号必须为数字");
         if (departmentMapper.findById(Integer.parseInt(id)) == null) throw new ParamException("该科室不存在");
         return departmentMapper.deleteById(Integer.parseInt(id));
     }
 
+    @Override
+    public List<ListVo> findAll(int uid, int hid, String name) throws DataException {
+        Map<String, String> map = authDataService.getAuthDataByUid(uid);
+        if (map.size() == 0) throw new DataException("当前用户没有数据权限,请联系管理员");
+        if (!map.containsKey(CoreConfig.AUTH_DATA_ALL)) throw new DataException("当前用户无最高数据权限，暂无法查看数据!");
+        return departmentMapper.findAll(hid, name);
+    }
+
     @Autowired
-    public DepartmentServiceImpl(DepartmentMapper departmentMapper, MapperFactory mapperFactory, HospitalMapper hospitalMapper) {
+    public DepartmentServiceImpl(DepartmentMapper departmentMapper, MapperFactory mapperFactory, HospitalMapper hospitalMapper, AuthDataService authDataService) {
         this.departmentMapper = departmentMapper;
         this.mapperFactory = mapperFactory;
         this.hospitalMapper = hospitalMapper;
+        this.authDataService = authDataService;
     }
 
     @Override
@@ -45,23 +64,16 @@ public class DepartmentServiceImpl implements DepartmentService {
         return departmentMapper.findOidByHid(hid);
     }
 
+    //ToDo:待完善
     @Override
-    public List<SelectVO> getDepartmentList(String name) {
-        return departmentMapper.getDepartmentList(name);
-    }
-
-    @Override
-    public List<SelectVO> getListByHid(String hid) {
-        return departmentMapper.getListByHid(hid);
-    }
-
-    @Override
-    public List<SelectVO> getListByHidOrName(String hid, String name) {
-        return departmentMapper.getListByHidOrName(hid, name);
-    }
-
-    @Override
-    public boolean insert(DepartmentVo departmentVo) throws ParamException {
+    public boolean insert(int uid, DepartmentVo departmentVo) throws ParamException, DataException {
+        Map<String, String> map = authDataService.getAuthDataByUid(uid);
+        if (map.size() == 0) throw new DataException("当前用户没有数据权限,请联系管理员");
+        if (!map.containsKey(CoreConfig.AUTH_DATA_ALL))
+            throw new DataException("当前用户没有最高数据权限，暂无法进行科室的添加操作!");
+        String[] strArray = map.get(CoreConfig.AUTH_DATA_HOSPITAL).split(",");
+        if (Arrays.stream(strArray).noneMatch(s -> s.equals(departmentVo.getHid())))
+            throw new DataException("当前用户没有医院权限,暂无法进行添加医院的操作");
         if (departmentVo.getHid() == null) throw new ParamException("请选择所属医院");
         Hospital hospital = hospitalMapper.findById(departmentVo.getHid());
         if (hospital == null) throw new ParamException("该医院不存在");
@@ -74,9 +86,13 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public boolean update(PutVo departmentPutVo) throws ParamException {
+    public boolean update(int uid, PutVo departmentPutVo) throws ParamException, DataException {
+        Map<String, String> map = authDataService.getAuthDataByUid(uid);
+        if (map.size() == 0) throw new DataException("当前用户没有数据权限,请联系管理员");
+        if (!map.containsKey(CoreConfig.AUTH_DATA_ALL))
+            throw new DataException("当前用户没有最高数据权限，暂无法进行科室的修改操作!");
         if (departmentPutVo.getId() == null) throw new ParamException("科室编号不能为空");
-        Department department=departmentMapper.findById(departmentPutVo.getId());
+        Department department = departmentMapper.findById(departmentPutVo.getId());
         if (department == null) throw new ParamException("该科室不存在");
         if (StringUtil.isEmpty(departmentPutVo.getName())) throw new ParamException("科室名称不能为空");
         if (departmentPutVo.getHid() == null) throw new ParamException("请选择所属医院");
